Finished stage #15 Added handleCreateLink to service
