+++ src/services/index.js
@@ -1,2 +1,3 @@
 export * from './userService'
 export * from './validateCreate'
+export * from './forumQuestions'
